Remove authentication by session_id to keep only authentication by JWT
