+++ routes/users.js
@@ -6,4 +6,14 @@ router.get('/', function(req, res, next) {
   res.send('respond with a resource');
 });
 
+/* Show all users */
+
+/* Show current user */
+
+/* Create new user */
+
+/* Update user */
+
+/* Delete user */
+
 module.exports = router;
